Avoid problems due to sub-seconds timestamps differences on export
Sometimes the filesytem would update a file's mtime with subseconds, but
the app-level comparison discarded the subsecond information and caused
and error when selecting the zonefiles to remove.

To avoid this problem, we now use a separate file to "touch" the files
when generating the zonefiles. Additionally, we explicitly set the
'change' timestamp in the unit test.