Replace 'domains#type' col with 'authority_type' and 'addressing_type'
The previous method of differentiating reverse and normal domains using
the 'name' column does not scale (string comparisons are expensive,
and we woudn't be able to index the column the way we should). So I've
added a 'addressing_type' column and, in the mean time, replaced the
ambiguous 'type' column with 'authority_type'. So now we have two domain
classifications: by address type (REVERSE/NORMAL) and by authority
type (MASTER/SLAVE).

Additionally, I've set the 'ttl' column as nullable, since slave domains
do not need a TTL.