Revert and fix e4118640
When named.conf had no changes after an 'import', we'd check whether any
changes were available with 'git add --verbose' and only added a new
revision to the repository if 'git add' returned something. However,
this breaks future exports which depend on the timestamp of the last
revision, which should be the timestamp of the 'import' operation.

Therefore, we now skip the 'git add' check and force a new revision of
the 'named.conf' file by adding a comment at the end with the 'import'
timestamp.