Exclude "session.key" from the importer's rsync
Since we're running the remote copies using the "dnsapi" user, we must now
exclude the "session.key" file. The reason is that it is created by the BIND
application itself, using the "named" user, and the permissions are set as 600
(so the "dnsapi" user can't access it and the rsync command will fail if we do
not explicitly exclude it).