Exclude IPs in "match-clients" to force matching through "keys"
The way Globo is attempting to use BIND views complicates the way we
must setup the servers to ensure that we're able to reach both views on
the master and slave servers. More especifically, BIND views are
supposed to use subnets that have no intersections -- i.e.
192.168.1.0/24 and 192.168.2.0/24. When the two subnets intersect, BIND
simply selects the first matching view. This is, supposedly, to simplify the
primary use case for views: defining a view for your intranet (i.e,
10.0.0.0/8), and another for everybody else (0.0.0.0/0).

Now, as stated initially, Globo is attempting to use views in a very
different way: to separate ipv4 from ipv6 addresses. The issue is that
the definition of the subnets in this case (0.0.0.0/0 and ::0/0) do not
create any actual subnets (both encompass the entire address spaces).
This means that the view selection will happen based on the type of the
IP address used to issue the query. *But*, if host originating the query
doesn't have an IP address of each type, it becomes impossible to query
for the view of the other type. An example might help clear things a
bit: on the QA environment, no server has an ipv6 address; so, there's
no way for the dnsapi server to issue queries on the domains/records
defined in the V6 view, nor for the slave server to transfers the
zones from the same view.

After a lot of reading and testing, I believe there are two solutions:
the first one -- which I skipped because I thought might be a bit too
limiting, is to require all servers to have an ipv6 address, set the
"transfer-source" statement accordingly in the definition of the slave
zones, and use the "-b" flag of the "dig" command to select the source
IP used to send the queries from the dnsapi server.

The second one, which is what this patch partially implements, is to use
the "key" statement to select the view. The "partially" is because the
patch actually augments the solution I've originally used to select the
views when implementing the "query" functionality long ago. At the time,
however, the tests didn't account for the fact the dnsapi server's IP
address could match the values from the "match-clients" statement of
both views. So, now, what we do is to explicitly remove the address of
the dnsapi and slave servers in the "match-clients" statement. And,
additionally, we explicitly set the "key" that should be used when
defining the "master" statement of a slave zone.

My preliminary tests in the QA environment seem to have worked well, but
this setup seems a bit fragile and it's absolutely possible I have
overlooked one or more use cases that will break under this setup.