Adding support to more than one slave
the config/globodns.yml file now sets a master dns server and a list of slaves.

All the code related on importing/exporting and interface has been
changed to comply to more than one slave.

The export strategy may still be clumsy and it is still going to be
polished.