Revert remote config when export fails.
Now when the export fails, the local git repo is reset, and
this old config is exported again.

This is important, because previously when we had an error, the local
repo would be reset, but the bind server would be stuck with falsy
config.

It is important to note that we test with rndc-checkconf before
syncing with the server. Thus, if we have any problem locally, we
can just reset the git repo, since no info was sync to server.

The server info is sync to the last valid commit ONLY IF anything
was already sent on this export. This happens if we have an excetion
AFTER the sync_remote method call.