Support for sibling domains added.
Suppose you have the same domain on more than one view,
and that these views share some records.

On legacy code, if you got to edit the domain inside one view,
the domain on the other view would be unchanged, and you would
have an inconsistency. The fix was to always remember to do the
same change on all 'sibling' domains.

Now, when we import, we check if there's another domain already
saved with the same name. If that's the case, we 'merge' the records
of both.

This merging happen in two steps: firstly, we remove all
replicated records of the new domain. Secondly, we set the
former domain as the 'sibling' of this new one. This relation
persists forever, and can be seen on the edit page for this particular
domain.

When we export again, we rebuild the domains as if they were
independent, but now without inconsistency.