Fixing bugs regarding slave export
The previous refactoring for more than one slave brought some
bugs because it tried to export zonefiles for slaves, before
creating them.

The original behavior of the system is to not export zonefiles
for slaves. This behavior was therefore preserved.

Another mistake was regarding disabling slaves on the conf file.
It is stated that just putting an empty ("") host for a slave
is enough for disabling it. This behavior wasn't persisted
on the cited refactoring and was now restored.

The last problem was on the revert method, that is supposed to
rollback the local git server and sync that version to the remote
servers. There was a problem when an error ocurred before the
commit/sync. That way the system tried to revert 'nothing'. Fixed.