Domain create api now supports passing a view.
You may pass the view id or name on the post call.
If you don't pass any of these, it is created with no view associated.
If you pass any of these, but the system can't find such a view, an
error is returned.

The method has been refactored as well, and a bug (500 when a falsy
template is passed) was automatically fixed.