Adding support to view on domain template
The domain template actually already had the view_id field, but
it wasn't being used.

The field was properly set on the model and on the build method
(that applies it to the domain being built)

UI has been updated with the ajax handlers to edit such field as
well.